pyproject.toml
[project]
name = "smartestate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=2.0.8",
    "easyocr>=1.7.2",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pycocotools>=2.0.10",
    "pytesseract>=0.3.13",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.34",
    "psycopg[binary]>=3.2.1",
    "elasticsearch>=8.12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.1.1",
    "pypdf>=4.3.1",
    "requests>=2.32.3",
    "tenacity>=9.0.0",
    "python-multipart>=0.0.9",
    "pytest>=8.3.3",
    "langchain>=0.3.2",
    "langgraph>=0.2.28",
    "langchain-community>=0.3.1",
    "langchain-elasticsearch>=0.3.0",
    "reportlab>=4.2.2",
    "httpx>=0.27.2",
    "jinja2>=3.1.4",
    "streamlit>=1.39.0",
]
